% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{profinit_cols}
\alias{profinit_cols}
\title{Profinit colors}
\usage{
profinit_cols(..., named = FALSE)
}
\arguments{
\item{...}{Either character name(s) or order of Profinit colors.
Leave empty for the full color set.}

\item{named}{Flag whether to produce named vector of profinit colours.
Defautls to \code{FALSE}.}
}
\value{
Named character vector the same length as input params.
}
\description{
This function provides an access to standardized Profinit colors hex codes.
}
\examples{
# to get Profinit red hexcode
profinit_cols("red")

# to get Profinit red and gray hexcodes
profinit_cols("red", "grey")

# to get the first Profinit color
profinit_cols(1)

# to get all Profinit colors
profinit_cols()
scales::show_col(profinit_cols())

}
