% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{profinit_pal.pals}
\alias{profinit_pal.pals}
\title{List of available Profinit palettes}
\usage{
profinit_pal.pals()
}
\value{
A character vector of all available palettes in this package.
}
\description{
Note: Function name has been chosen to match the \code{grDevices::palette.pals()}.
}
\examples{
profinit_pal.pals()

# Now you are able to use this information, e.g.:
reds_palette <- profinit_pal("reds")
reds_palette(4)
scales::show_col(reds_palette(4))

}
