% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiplots-package.R
\docType{package}
\name{profiplots-package}
\alias{profiplots}
\alias{profiplots-package}
\title{profiplots: Profinit Plotting Theme}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Help unify visual output of R analyses in the Profinit EU company. So far, there are color and fill scales for 'ggplot2', plotting theme for 'ggplot2', color palettes and utils to make the tools default choices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://profinit.eu}
}

}
\author{
\strong{Maintainer}: Dominik Matula \email{dominik.matula@profinit.eu}

Other contributors:
\itemize{
  \item Profinit EU, Ltd. [copyright holder, funder]
}

}
\keyword{internal}
